requirements.txt
Flask==2.0.1
Flask_Cors==3.0.10
pycryptodome==3.17
requests==2.25.1
waitress==2.1.2
